refactor: 자주 사용되는 Http 헤더 키 enum 처리
